refactor: update substrait.gen.proto imports to just substrait
