+++ src/substrait/builders/type.py
@@ -1,6 +1,6 @@
 from typing import Iterable
 
-import substrait.gen.proto.type_pb2 as stt
+import substrait.type_pb2 as stt
 
 
 def boolean(nullable=True) -> stt.Type:
