+++ src/substrait/extension_registry/function_entry.py
@@ -3,9 +3,10 @@
 from enum import Enum
 from typing import Optional, Union
 
+from substrait.type_pb2 import Type
+
 from substrait.derivation_expression import _parse, evaluate
 from substrait.gen.json import simple_extensions as se
-from substrait.gen.proto.type_pb2 import Type
 
 from .signature_checker_helpers import covers, normalize_substrait_type_names
