+++ src/substrait/extension_registry/signature_checker_helpers.py
@@ -2,9 +2,10 @@
 
 from typing import Dict
 
+from substrait.type_pb2 import Type
+
 from substrait.derivation_expression import _evaluate
 from substrait.gen.antlr.SubstraitTypeParser import SubstraitTypeParser
-from substrait.gen.proto.type_pb2 import Type
 
 from .exceptions import UnhandledParameterizedTypeError, UnrecognizedSubstraitTypeError
