+++ tests/builders/extended_expression/test_column.py
@@ -1,6 +1,7 @@
-import substrait.gen.proto.algebra_pb2 as stalg
-import substrait.gen.proto.extended_expression_pb2 as stee
-import substrait.gen.proto.type_pb2 as stt
+import substrait.algebra_pb2 as stalg
+import substrait.extended_expression_pb2 as stee
+import substrait.type_pb2 as stt
+
 from substrait.builders.extended_expression import column
 
 struct = stt.Type.Struct(
