+++ tests/builders/extended_expression/test_window_function.py
@@ -1,9 +1,9 @@
+import substrait.algebra_pb2 as stalg
+import substrait.extended_expression_pb2 as stee
+import substrait.extensions.extensions_pb2 as ste
+import substrait.type_pb2 as stt
 import yaml
 
-import substrait.gen.proto.algebra_pb2 as stalg
-import substrait.gen.proto.extended_expression_pb2 as stee
-import substrait.gen.proto.extensions.extensions_pb2 as ste
-import substrait.gen.proto.type_pb2 as stt
 from substrait.builders.extended_expression import window_function
 from substrait.extension_registry import ExtensionRegistry
