+++ tests/builders/plan/test_aggregate.py
@@ -1,9 +1,9 @@
+import substrait.algebra_pb2 as stalg
+import substrait.extensions.extensions_pb2 as ste
+import substrait.plan_pb2 as stp
+import substrait.type_pb2 as stt
 import yaml
 
-import substrait.gen.proto.algebra_pb2 as stalg
-import substrait.gen.proto.extensions.extensions_pb2 as ste
-import substrait.gen.proto.plan_pb2 as stp
-import substrait.gen.proto.type_pb2 as stt
 from substrait.builders.extended_expression import aggregate_function, column
 from substrait.builders.plan import aggregate, default_version, read_named_table
 from substrait.builders.type import boolean, i64
