+++ tests/builders/plan/test_read.py
@@ -1,13 +1,13 @@
 import pytest
+import substrait.algebra_pb2 as stalg
+import substrait.plan_pb2 as stp
+import substrait.type_pb2 as stt
 from google.protobuf import any_pb2
 from google.protobuf.wrappers_pb2 import StringValue
+from substrait.extensions.extensions_pb2 import AdvancedExtension
 
-import substrait.gen.proto.algebra_pb2 as stalg
-import substrait.gen.proto.plan_pb2 as stp
-import substrait.gen.proto.type_pb2 as stt
 from substrait.builders.plan import default_version, read_named_table
 from substrait.builders.type import boolean, i64
-from substrait.gen.proto.extensions.extensions_pb2 import AdvancedExtension
 
 struct = stt.Type.Struct(
     types=[i64(nullable=False), boolean()],
