+++ tests/extension_registry/test_registry_lookup.py
@@ -1,5 +1,6 @@
+from substrait.type_pb2 import Type
+
 from substrait.builders.type import decimal, i8, i16
-from substrait.gen.proto.type_pb2 import Type
 
 
 def test_non_existing_urn(registry):
