+++ tests/test_proto.py
@@ -3,15 +3,15 @@
 
 def test_imports():
     """Temporary sanity test"""
-    from substrait.gen.proto.algebra_pb2 import Expression
-    from substrait.gen.proto.capabilities_pb2 import Capabilities
-    from substrait.gen.proto.extended_expression_pb2 import ExtendedExpression
-    from substrait.gen.proto.extensions.extensions_pb2 import SimpleExtensionURN
-    from substrait.gen.proto.function_pb2 import FunctionSignature
-    from substrait.gen.proto.parameterized_types_pb2 import ParameterizedType
-    from substrait.gen.proto.plan_pb2 import Plan
-    from substrait.gen.proto.type_expressions_pb2 import DerivationExpression
-    from substrait.gen.proto.type_pb2 import Type
+    from substrait.algebra_pb2 import Expression
+    from substrait.capabilities_pb2 import Capabilities
+    from substrait.extended_expression_pb2 import ExtendedExpression
+    from substrait.extensions.extensions_pb2 import SimpleExtensionURN
+    from substrait.function_pb2 import FunctionSignature
+    from substrait.parameterized_types_pb2 import ParameterizedType
+    from substrait.plan_pb2 import Plan
+    from substrait.type_expressions_pb2 import DerivationExpression
+    from substrait.type_pb2 import Type
 
 
 def test_proto_proxy_module():
