+++ tests/test_type_inference.py
@@ -1,5 +1,6 @@
-import substrait.gen.proto.algebra_pb2 as stalg
-import substrait.gen.proto.type_pb2 as stt
+import substrait.algebra_pb2 as stalg
+import substrait.type_pb2 as stt
+
 from substrait.type_inference import (
     infer_expression_type,
     infer_nested_type,
